psQTL_prep major changes; VCF filtering
Other than the filtering being added in, and the caches needing to adjust to accommodate that, I've refactored how VCF files are parsed.

Note that at this point, I am horrified by how the caches are being handled. I'd like to refactor that system entirely but I've only just managed to get it working and there's more important things to do. If I find more bugs in the future, it must be replaced wholesale.