psQTL_prep try to mitigate ACID issues
If we're running a 'depth' and 'call' analysis at the same time, we might end up ignoring parameters set in one run in favour of whichever one finishes last. By reloading the cache in whenever we store a new output, we'll try to reduce that risk. With this implementation, the only threat is that the file gets opened and modifying at the exact same time in which case we will have an issue with isolation.