psQTL_post coverage plot WIP
I'm not sure if this is going to turn out well. I think it's pushing the user experience over the line from "a bit intimidating, but not too bad" to "there's too many parameters that sound the same" which I don't think should be tolerated unless these plots really help to elucidate trends in the data enough to warrant their inclusion.