Major code shuffling
Lots of stuff in plotting.py is being put into the generalised Plot class which will focus on data formatting. That Plot class is intended to be inherited and built upon by HorizontalPlot (which will do basically what psQTL is currently doing) and CircosPlot (which will give the alternate circos visualisation) as classes implementing concrete logic for visualisation.