Minor refactor of Plot class
Plot classes store the data directly in order for it to know what it is going to plot. Makes it easier for better legend plotting when the refactor extends to the CircosPlot class. Problems with this commit: SNP ED4 is different, and gene model aesthetics are weird (exons not plotted?)