Fix line ending extension and smoothing
Using a buffer then truncating things based on a projection of the slope gives a "true" idea of how the data would be plotted if it were to go off plot. This was done implicitly by the horizontal plot plotting "off screen" then narrowing the xlim, but for Circos plotting that isn't allowed.