Fixed line interpolation; stick selected to line
Selected SNPs/CNVs do not have a Y-value that matches the BA axis of the track they reside on. It makes sense to me that these selected features should stick to the line for clarity of interpretation. Otherwise, fixed a bug with how data interpolation was being performed w/r/t extending the line to the edges of the plot.