Implemented inheritance ED and associated tests
Updated the ED formula Excel sheet to allow playing with how it works. Code should allow this form of ED to work in lieu of the existing formula when parent samples are provided, but this hasn't been practically tested yet. Also, I would like to allow both types of results to coexist and to choose which to plot later if both exist.