v1.2.1; polyploidy handling, better sPLS-DA handling
Modifying code behaviour to allow polyploidy occurred in ed.py with a simple tweak. sPLS-DA recoding now allows for data to be filtered by metadata mismatch to original VCF, which conforms with the rest of psQTL's behaviour. Also identified issues with sPLS-DA when only one variant is located and made it generally much more robust. Lots of extra tests of whole-program behaviour for validation of robustness.