v1.2.2; overhaul to ED calculation
The previous method of calculating ED is not leveraging one of the main benefits of actually performing per-sample sequencing and segregation from those genotype calls -- we can properly differentiate heterozygotes from a mix of homozygotes. The update removes the use of the modified ED formula since it shouldn't hold any benefit now.