v1.2.4: revert back to previous ED calculation
The new ED calculation, despite making sense in theory, has problems when working with real (and messy) data. I'm not entirely certain why it works so poorly, but suspect that it relates to spurious read mapping and other data-level issues. Reversion to the previous approach reinstates a more robust and tested algorithm.