v1.2.5; polyploid calling out of scope
This version increment denotes the intentional choice to make clear how psQTL will not (at this time) attempt to internally call polyploid variants, despite all other systems supporting that.

I may in the future to decide to implement a simple GATK-based system for polyploids, but for now it's likely better for someone to call their own variants using a system specifically tuned for their organism.