Produce all Euclidean distances we can
Produces allele frequency, genotype frequency, and inheritance ED values if parent genotypes are provided; the first two only if not. The first two are left unfiltered intentionally to allow for naive investigation of results without considering parents. The inheritance ED is where we want filtering to occur for a more nuanced view of potential issues in the data.