windowed_splsda more error handling
This commit attempts to clarify some errors which make sense in the context of the R code but are gibberish to a front end psQTL user. Changes the metadata file parsing to not mandate tab-separation. Additionally added in more specific error detection for incompatibilities between the metadata and the encoded VCF.