windowed_splsda guardrails for R column labels
R modifies the encoded VCF's column/sample labels when loading in as a dataframe, which can cause incompatibility with the metadata table. Metadata labels are adjusted to accommodate this now and prevent those errors. I may want to relabel the samples when outputting to table but that is a low priority concern.